MAINT: canonicalize processing.alignment namespace
