+++ src/spectrochempy/processing/alignment/__init__.py
@@ -4,18 +4,9 @@
 # See full LICENSE agreement in the root directory.
 # ======================================================================================
 
-import warnings
-
 import lazy_loader as _lazy_loader
 
 # --------------------------------------------------------------------------------------
 # Lazy loading of sub-packages
 # --------------------------------------------------------------------------------------
 __getattr__, __dir__, __all__ = _lazy_loader.attach_stub(__name__, __file__)
-
-warnings.warn(
-    "Importing from `spectrochempy.processing.alignement` is deprecated and "
-    "will be removed in 0.11.0. Use `spectrochempy.processing.alignment` instead.",
-    DeprecationWarning,
-    stacklevel=2,
-)
